# docs/swagger.yaml
basePath: /v1
definitions:
  domainpayment.PaymentType:
    enum:
    - CASH
    - E-WALLET
    - EDC
    type: string
    x-enum-varnames:
    - Cash
    - EWallet
    - EDC
...
  modelv1.CreatePaymentRequest:
    properties:
      logo:
        example: https://example.com/cash.png
        type: string
      name:
        example: Tunai
        type: string
      type:
        allOf:
        - $ref: '#/definitions/domainpayment.PaymentType'
        example: CASH
    required:
    - logo
    - name
    - type
    type: object
...
  modelv1.ErrorResponse:
    properties:
      messages:
        example:
        - Error message 1
        - ' Error message 2'
        items:
          type: string
        type: array
      success:
        example: false
        type: boolean
    type: object
...
  modelv1.Meta:
    properties:
      limit:
        example: 10
        type: integer
      skip:
        example: 0
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  modelv1.PaymentResponse:
    properties:
      id:
        example: 1
        type: integer
      logo:
        example: https://example.com/cash.png
        type: string
      name:
        example: Tunai
        type: string
      type:
        allOf:
        - $ref: '#/definitions/domainpayment.PaymentType'
        example: CASH
    type: object
...
  modelv1.Response:
    properties:
      data: {}
      message:
        example: Success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  modelv1.UpdatePaymentRequest:
    properties:
      logo:
        example: https://example.com/gopay.png
        type: string
      name:
        example: Gopay
        type: string
      type:
        allOf:
        - $ref: '#/definitions/domainpayment.PaymentType'
        example: E-WALLET
    required:
    - logo
    - name
    - type
    type: object
...
host: localhost
info:
  contact:
    email: letienminh2512@gmail.com
    name: Tien Minh
    url: https://github.com/TienMinh25/go-hexagonal-architecture
  description: This is a simple RESTful Point of Sale (POS) Service API written in
    Go using Gin web framework, PostgreSQL database, and Redis cache.
  license:
    name: APACHE
    url: https://github.com/TienMinh25/go-hexagonal-architecture/blob/main/LICENSE
  title: Go Sale API demo for hexagonal architecture
  version: "1.0"
paths:
...
  /payments:
    get:
      consumes:
      - application/json
      description: List payments with pagination
      parameters:
      - description: Skip
        in: query
        name: skip
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Payments displayed
          schema:
            $ref: '#/definitions/modelv1.Meta'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List payments
      tags:
      - Payments
    post:
      consumes:
      - application/json
      description: create a new payment with name, type, and logo
      parameters:
      - description: Create payment request
        in: body
        name: createPaymentRequest
        required: true
        schema:
          $ref: '#/definitions/modelv1.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Payment created
          schema:
            $ref: '#/definitions/modelv1.PaymentResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new payment
      tags:
      - Payments
  /payments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a payment by id
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Payment deleted
          schema:
            $ref: '#/definitions/modelv1.Response'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a payment
      tags:
      - Payments
    get:
      consumes:
      - application/json
      description: get a payment by id
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Payment retrieved
          schema:
            $ref: '#/definitions/modelv1.PaymentResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a payment
      tags:
      - Payments
    put:
      consumes:
      - application/json
      description: update a payment's name, type, or logo by id
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update payment request
        in: body
        name: updatePaymentRequest
        required: true
        schema:
          $ref: '#/definitions/modelv1.UpdatePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Payment updated
          schema:
            $ref: '#/definitions/modelv1.PaymentResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a payment
      tags:
      - Payments
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"